requirements.txt
# ===============================
# Core Flask stack
# ===============================
Flask==3.0.3
Werkzeug==3.0.1

# ===============================
# Database & ORM
# ===============================
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0

# ===============================
# Authentication & Forms
# ===============================
Flask-Login==0.6.3
Flask-WTF==1.2.2
email-validator==2.2.0

# ===============================
# Security
# ===============================
cryptography==43.0.0

# ===============================
# Timezone support (IMPORTANT for Windows)
# ===============================
tzdata==2024.1

# ===============================
# OPTIONAL but RECOMMENDED
# ===============================
Flask-Migrate==4.0.7
python-dotenv==1.0.1
